Refactoring and small changes
