+++ src/main/resources/templates/professor/lessons/content.html
@@ -16,10 +16,10 @@
                 <!-- Title -->
                 <p th:text="${lesson.getTitle()}" class="et-name text-2xl font-bold">Lesson name</p>
                 <!-- Lesson content -->
-                <p th:text="${lesson.getContent()}" class="et-period text-xl">This lesson is so interesting</p>
+                <p th:text="${lesson.getContent()}" class="et-description mt-4">This lesson is so interesting</p>
             </div>
             <!-- Modify button is disabled for students-->
-                <button th:if="${canEdit}" id="btn-edit-lesson-from-content"
+                <button th:if="${canEdit}" id="btn-edit-lesson"
                         th:onclick="'document.location=\'/professor/courses/'+${lesson.getCourse().getId()}+'/lessons/'+${lesson.getId()}+'/edit\''"
                         class="et-button-primary w-28">Modify
                 </button>
