+++ src/test/java/it/ettore/e2e/Login.java
@@ -1,8 +1,6 @@
 package it.ettore.e2e;
 
-import it.ettore.e2e.po.professor.ProfessorCoursesPage;
 import it.ettore.e2e.po.LoginPage;
-import it.ettore.e2e.po.RegisterPage;
 import it.ettore.model.User;
 import it.ettore.model.UserRepository;
 import org.junit.Test;
