+++ src/test/java/it/ettore/e2e/po/LoginPage.java
@@ -1,6 +1,6 @@
 package it.ettore.e2e.po;
 
-import it.ettore.e2e.po.professor.ProfessorCoursesPage;
+import it.ettore.e2e.po.professor.courses.ProfessorCoursesPage;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.WebElement;
