+++ src/test/java/it/ettore/e2e/professor/courses/ProfessorCourse.java
@@ -2,7 +2,7 @@ package it.ettore.e2e.professor.courses;
 
 import it.ettore.e2e.E2EBaseTest;
 import it.ettore.e2e.po.LoginPage;
-import it.ettore.e2e.po.professor.ProfessorCoursePage;
+import it.ettore.e2e.po.professor.courses.ProfessorCoursePage;
 import it.ettore.model.Course;
 import it.ettore.model.CourseRepository;
 import it.ettore.model.User;
