+++ src/test/java/it/ettore/e2e/professor/courses/ProfessorCourses.java
@@ -2,8 +2,8 @@ package it.ettore.e2e.professor.courses;
 
 import it.ettore.e2e.E2EBaseTest;
 import it.ettore.e2e.po.LoginPage;
-import it.ettore.e2e.po.professor.ProfessorCoursesPage;
-import it.ettore.e2e.po.professor.ProfessorCoursesPage.CourseComponent;
+import it.ettore.e2e.po.professor.courses.ProfessorCoursesPage;
+import it.ettore.e2e.po.professor.courses.ProfessorCoursesPage.CourseComponent;
 import it.ettore.model.Course;
 import it.ettore.model.CourseRepository;
 import it.ettore.model.User;
